// src/store/modules/movies/action-types.ts
/* eslint-disable no-shadow */

export enum MoviesActionTypes {
  GET_MOVIES_BY_QUERY = 'GET_MOVIES_BY_QUERY',
  ADD_FILTER_OPTION = 'ADD_FILTER_OPTION',
  REMOVE_FILTER_OPTION = 'REMOVE_FILTER_OPTION',
  SET_SEARCH_QUERY = 'SET_SEARCH_QUERY',
  SET_SEARCH_PARAM = 'SET_SEARCH_PARAM',
  SET_CURRENT_PAGE = 'SET_CURRENT_PAGE',
  SET_MOVIES_PER_PAGE = 'SET_MOVIES_PER_PAGE',
}